handle illegal mongodb keys

+++ lib/puppetdb_query/mongodb.rb
@@ -2,6 +2,18 @@ require 'time'
 
 require_relative "logging"
 
+# monkey patch mongodb to get rid of
+# ".. is an illegal key in MongoDB. Keys may not start with '$' or contain a '.'. (BSON::String::IllegalKey)"
+module Mongo
+  module Protocol
+    class Message
+      def validating_keys?
+        false
+      end
+    end
+  end
+end
+
 module PuppetDBQuery
   # access nodes and their facts from mongo database
   # rubocop:disable Metrics/ClassLength
@@ -151,17 +163,27 @@ module PuppetDBQuery
     # update or insert facts for given node name
     #
     # @param node [String] node name
-    # @param facts [Array<String>] get these facts in the result, eg ['fqdn'], empty for all
+    # @param facts [Hash] facts for the node
     def node_update(node, facts)
-      connection[nodes_collection].find(_id: node).replace_one(facts, upsert: true, bypass_document_validation: false,  check_keys: false)
+      logger.debug "  updating #{node}"
+      connection[nodes_collection].find(_id: node).replace_one(facts, upsert: true,
+        bypass_document_validation: true, check_keys: false, validating_keys: false)
     rescue ::Mongo::Error::OperationFailure => e
+      logger.warn "  updating #{node} failed with: #{e.message}"
       # mongodb doesn't support keys with a dot
       # see https://docs.mongodb.com/manual/reference/limits/#Restrictions-on-Field-Names
       # as a dirty workaround we delete the document and insert it ;-)
       # The dotted field .. in .. is not valid for storage. (57)
-      raise e unless e.message =~ /The dotted field /
-      connection[nodes_collection].find(_id: node).delete_one
-      connection[nodes_collection].insert_one(facts.merge(_id: node), check_keys: false)
+      # .. is an illegal key in MongoDB. Keys may not start with '$' or contain a '.'. (BSON::String::IllegalKey)
+      raise e unless (e.message =~ /The dotted field / || e.message =~ /is an illegal key/)
+      logger.warn "    we try again deleting and inserting the node"
+      begin
+        connection[nodes_collection].find(_id: node).delete_one
+        connection[nodes_collection].insert_one(facts.merge(_id: node),
+          bypass_document_validation: true, check_keys: false, validating_keys: false)
+      rescue
+        logger.error "  updating #{node} failed with: #{e.message}"
+      end
     end
 
     # delete node data for given node name
